+++ packages/field/src/components/TextArea/readonly.tsx
@@ -11,7 +11,7 @@ import type { ProFieldFC } from '../../index';
  */
 const FieldTextAreaReadonly: ProFieldFC<{
   text: string;
-}> = ({ text }, ref) => {
+}> = ({ text, fieldProps }, ref) => {
   const { getPrefixCls } = useContext(ConfigProvider.ConfigContext);
   const readonlyClassName = getPrefixCls('pro-field-readonly');
   const compClassName = `${readonlyClassName}-textarea`;
@@ -20,7 +20,6 @@ const FieldTextAreaReadonly: ProFieldFC<{
     return {
       [`.${compClassName}`]: {
         display: 'inline-block',
-        // padding: '4px 11px',
         lineHeight: '1.5715',
         maxWidth: '100%',
         whiteSpace: 'pre-wrap',
@@ -32,7 +31,7 @@ const FieldTextAreaReadonly: ProFieldFC<{
     <span
       ref={ref}
       className={classNames(hashId, readonlyClassName, compClassName)}
-      style={{}}
+      {...fieldProps}
     >
       {text ?? '-'}
     </span>,