+++ packages/form/src/index.tsx
@@ -15,6 +15,7 @@ export type {
   FormListActionType,
   ProFormCaptchaProps,
   ProFormCheckboxGroupProps,
+  ProFormCheckboxProps,
   ProFormColorPickerProps,
   ProFormColumnsType,
   ProFormDependencyProps,