+++ packages/form/CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.27.0](https://github.com/ant-design/pro-components/compare/@ant-design/pro-form@2.26.5...@ant-design/pro-form@2.27.0) (2024-05-20)
+
+### Features
+
+- add onBlur to filedChildren ([#8418](https://github.com/ant-design/pro-components/issues/8418)) ([3b104b9](https://github.com/ant-design/pro-components/commit/3b104b9e590e1eef139a5164bffc70b243389db3))
+
+## [2.26.6](https://github.com/ant-design/pro-components/compare/@ant-design/pro-form@2.26.5...@ant-design/pro-form@2.26.6) (2024-04-21)
+
+**Note:** Version bump only for package @ant-design/pro-form
+
+## [2.26.5](https://github.com/ant-design/pro-components/compare/@ant-design/pro-form@2.26.4...@ant-design/pro-form@2.26.5) (2024-04-21)
+
+**Note:** Version bump only for package @ant-design/pro-form
+
 ## [2.26.4](https://github.com/ant-design/pro-components/compare/@ant-design/pro-form@2.26.2...@ant-design/pro-form@2.26.4) (2024-04-20)
 
 **Note:** Version bump only for package @ant-design/pro-form