+++ packages/utils/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.15.13](https://github.com/ant-design/pro-components/compare/@ant-design/pro-utils@2.15.11...@ant-design/pro-utils@2.15.13) (2024-05-20)
+
+**Note:** Version bump only for package @ant-design/pro-utils
+
+## [2.15.12](https://github.com/ant-design/pro-components/compare/@ant-design/pro-utils@2.15.11...@ant-design/pro-utils@2.15.12) (2024-04-21)
+
+**Note:** Version bump only for package @ant-design/pro-utils
+
+## [2.15.11](https://github.com/ant-design/pro-components/compare/@ant-design/pro-utils@2.15.10...@ant-design/pro-utils@2.15.11) (2024-04-21)
+
+**Note:** Version bump only for package @ant-design/pro-utils
+
 ## [2.15.10](https://github.com/ant-design/pro-components/compare/@ant-design/pro-utils@2.15.8...@ant-design/pro-utils@2.15.10) (2024-04-20)
 
 **Note:** Version bump only for package @ant-design/pro-utils